#19 - Test classic implementation of class vcl::Vector
Completed tests, fixed bugs.
Notice: all fromer .ixx modules have been removed also.
