--- vcl/tests/test_main.cpp
@@ -1,10 +1,0 @@
-/** \brief main for tests. */
-
-#include <cstdio>
-
-int main()
-{
-    printf("VCL testing application\n");
-
-    return 0;
-}
